Utility methods for contextual responses
This commit adds three new utility methods.

`talkingToMe` returns true if `msg` is a direct message, or mentions the Slack instance by name or @reference.

`reply` will reply to a message in context.  If `msg` is part of a thread, it will send `text` to the thread.  Otherwise, it will send `text` to the same channel as `msg`.

`thread` will send `text` to the same thread as `msg`.  If `msg` is not part of a thread, `thread` will use `msg` as the parent.